feat: re-introduce constraints + minor refactoring
The constancy and vanishing constraints aren't "inherited" from the
TXN_DATA module. The only point in time when the REL_BLOCK_NUMBER are
"set" is during the TX_INIT / TX_SKIP sections.

They aren't set as much as they are sent to the eponymous columns in the
TXN_DATA module. However, outside of transaction/ rows they are
relatively unconstrained. The present constraints being some of these.