select `go-corset` version `1.1.1`
This updates `go-corset` to the latest release which removes the notion
of loobean, and replaces it with a clear distinction between logical
conditions and integer values.  In particular, this adds the operators
`==` and `!=` along with true notions of logical or and logical and.
This also includes relational operators (e.g. `<` and `<=`) which can be
used in constrant expresions (only).