fix for `projected-height`
This function is part of the lookup constraint `hub-into-wcp-for-sox`.
It computes a negative number for a given field which, previously, was
being hidden by the fact that the selector was not active (i.e. so -1 *
0 gave 0).  Since the selector multiplication is no longer present, this
value was remaining as -1.

The solution here is to put back a multiplication which ensures the
negative value does not arise in practice.