remove LONDON-specific ADD module
This removes the LONDON implementation of the ADD module which, for
backwards compatibility reasons, was retained in its original Corset
lisp form.  Instead, a single assembly implementation of the ADD module
is used for all forks.

The assembly ADD implementation is also updated to use the new fail
instruction, which allows us to ensure that only valid instances of the
function are included in a trace.