add zkasm implementation of GAS module
This puts in place a ZK assembly implementation of the GAS module.  This
module differs slightly from the original in that it is implemented as a
one-line function, meaning we have no need for the IOMF flag.
Furthermore, this module does not defer to the WCP module but, instead,
performs its comparisons directly.