add zkasm implementation of WCP module
This puts in place a zkasm implementation of the WCP module.  There is,
however, one outstanding issue: the change in signature for the WCP
module in cancun.

At this stage, it seems to me that we have two options: first (and
easiest) is to have a single wcp.zkasm file for all forks; secondly, we
could have two distinct files (e.g. wcp_london.zkasm and
wcp_cancun.zkasm), where only the latter supports the additional
functionality.