fix EXP implementation
This puts through a fix for the EXP implementation to properly support
cases where CDS < EBS.  In such cases, we need to fill all bytes in
between with 0.  In actual fact, what the module does now is to always
fill the bytes from CDS onwards with 0.  This is potentially somewhat
ineffecient in the case that EBS < CDS, since the filled bytes are
wasted.  However, it simplifies the asm code significantly.

This implementation has now been tested against all tests in ExpTest
(including those previously marked as @Disabled).