fixes for KOALABEAR_16
The bit_shr and bit_sar utility functions were not able to compile on
the KOALABEAR_16 fields due a known existing issue with the zkasm
register splitting algorithm.  Previously, a workaround was used which
worked only for the BLS12_377 field.  However, a workaround has now been
developed which works for both BLS12_377 and KOALABEAR_16.