feat: use max_log() instead of wcp.CT_MAX
This replaces the use of wcp.CT_MAX with a call to the max_log()
function in the rlputils module.  The goal of this is break the
dependency of rlputils on the wcp.CT_MAX column.  This is needed in order
to replace the wcp module with an assembly implementation (i.e. where
the CT_MAX column will not exist).