apply unsafe casts to lookups into zkasm
This uses a number of unsafe casts to workaround lookup issues into
zkasm modules.  This should perhaps only be a temporary fix, as the
casts are unsafe (i.e. the cast does not ensure the given values is in
the casted range and, instead, relies on that to be shown elsewhere).